Move `AddressFormatter` and `ContactNameProvider` to :ui:legacy module
