+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/messageview/DisplayRecipientsExtractor.kt
@@ -1,9 +1,9 @@
 package com.fsck.k9.ui.messageview
 
 import com.fsck.k9.Account
-import com.fsck.k9.helper.AddressFormatter
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.Message
+import com.fsck.k9.ui.helper.AddressFormatter
 
 /**
  * Extract recipient names from a message to display them in the message view.
