+++ app/ui/legacy/src/main/java/com/fsck/k9/view/MessageHeader.java
@@ -27,8 +27,8 @@ import com.fsck.k9.contacts.ContactPictureLoader;
 import com.fsck.k9.helper.ClipboardManager;
 import com.fsck.k9.helper.Contacts;
 import com.fsck.k9.helper.MessageHelper;
-import com.fsck.k9.helper.RealAddressFormatter;
-import com.fsck.k9.helper.RealContactNameProvider;
+import com.fsck.k9.ui.helper.RealAddressFormatter;
+import com.fsck.k9.ui.helper.RealContactNameProvider;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.mail.Flag;
 import com.fsck.k9.mail.Message;
